move demande-de-stage to pages folder
