+++ src/app/app-routing.module.ts
@@ -8,7 +8,7 @@ import { CandidatsPageComponent } from './pages/candidats-page/candidats-page.co
 import { PageLoginComponent } from './pages/page-login/page-login.component';
 import { CandidatAjoutPageComponent } from './pages/candidat-ajout-page/candidat-ajout-page.component';
 import { EntrepriseAjoutPageComponent } from './pages/entreprise-ajout-page/entreprise-ajout-page.component';
-import { DemandeDeStageComponent } from './demande-de-stage/demande-de-stage.component';
+import { DemandeDeStageComponent } from './pages/demande-de-stage/demande-de-stage.component';
 
 const routes: Routes = [
   { path: '', redirectTo: 'page-login', pathMatch: 'full' },
