+++ src/app/pages/demande-de-stage/demande-de-stage.component.scss
@@ -1,9 +1,9 @@
-@import "../../SCSS/vars";
+@import "../../../SCSS/vars";
 
 :host {
-    display: flex;
-    background-color: $gris-clair;
-    padding-bottom: 50px;
+  display: flex;
+  background-color: $gris-clair;
+  padding-bottom: 50px;
 }
 .container {
   max-width: 1140px;
@@ -12,9 +12,7 @@
   padding-right: 15px;
   padding-left: 15px;
   padding-top: 50px;
-  
 }
-
 
 .title {
   border-left: 8px solid $jaune-fonce;
@@ -24,7 +22,7 @@
 .title p {
   font-size: 14px;
   font-weight: bold;
-  color:$gris-fonce;
+  color: $gris-fonce;
 }
 
 .title h1 {
@@ -36,26 +34,23 @@
 .icons {
   display: flex;
   justify-content: flex-end;
-  
-  
 }
 
-.check-icon{
+.check-icon {
   color: $btn-accepter;
   transform: scale(1.5);
   margin-right: 20px;
 }
 
-.edit-icon{
+.edit-icon {
   color: $bleu-fonce;
   transform: scale(1.5);
   margin-right: 20px;
 }
 
-.close-icon{
+.close-icon {
   color: $rouge;
   transform: scale(1.5);
-  
 }
 
 .card-container {
@@ -87,8 +82,6 @@ h2 {
   border-left: 8px solid $gris-fonce;
   padding-left: 10px;
   width: 300px;
-  
-
 }
 
 .inner-div p {
@@ -103,9 +96,8 @@ h2 {
   color: $gris-fonce;
 }
 
-.inner-div-right{
-
-  margin-right: 20%
+.inner-div-right {
+  margin-right: 20%;
 }
 
 .section-title {
@@ -130,11 +122,9 @@ h2 {
   margin-top: 30px;
 }
 
-.download-icon{
-
+.download-icon {
   margin-right: 10px;
 }
-
 
 /* Media queries */
 @media only screen and (max-width: $w-medium) {
